requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
scikit-image==0.25.2
tk==0.1.0
pillow==11.3.0
onnx==1.18.0
onnxruntime-gpu==1.22.0
protobuf==6.31.1
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
tqdm
ftfy
regex==2025.7.34
flask